Add the type as a parameter to the eval_offset function
